Update README instructions
The README instructions now state how to modify the environment variables to work with local development, both for executing Docker and for building the backend/frontend without Docker.

Additionally, this commit changes the IP address that was being used for the Docker address during tests to the same address that is defined in the environment variables for consistency.